Simulation 5.0

+++ include/simulation/TrafficLight.hpp
@@ -12,7 +12,7 @@ class TrafficLight : public VisualElement
     private:
     std::vector<std::pair<StreetConnectionIDs, bool>> connections; //The second item of the pair is a bool that contains true if the light is green and false if the light is red
     int time_to_change;
-    int last_change;
+    float last_change;
 
     public: 
 
@@ -26,7 +26,7 @@ class TrafficLight : public VisualElement
     std::vector<std::pair<StreetConnectionIDs, bool>> get_connections();
     int get_time_to_change();
 
-    void update(int current_time);
+    void update(float current_time);
 
     nlohmann::json to_json();
 };
